Renewing the app theme mode management system
First of all, the old BLoC files of the theme were deleted, subsequently a folder called models was created where a file was created with the AppSettingsModel class which has within it the data representing the Theme Mode and also the copyWith method which allows us to alter data within the class. The files for BLoC were then created and the files where there were references to the old BLoC files were modified.